Rename cops moved from the Style department to Naming department in Rubocop v0.50.0
Rubocop adding a Naming department here:
https://github.com/bbatsov/rubocop/issues/4521

and now it complains with warnings if you use the deprecated names for
the cops, which breaks all sorts of editor integrations because they
can't parse out the warnings. Sigh. So, lets update our names to the new
ones! This won't work for old rubocop versions that don't have a Naming
department to find these cops under, but, those projects should be able
to upgrade rubocop and the problem should go away. Horray!