Add Style/MethodCallWithArgsParentheses
The main motivation is of course consistency. Now instead of deciding if the method should put
parentheses or now, we always put them, the exceptions are on class macros and some ignored methods.

Given the Gemfile is a heavy user of DSLs it is completely excluded.

This change cop also brings one advantage in terms of code safety that block precedence now
is easier to understand since you can use parentheses to define it. That brings to one problem
that excluding class macros causes, they are also suppressible to block precedence problems.