Create CODEOWNERS
This commit creates a `CODEOWNERS` file in the root of the repo and adds myself and Rafael França as global codeowners. This will go side-by-side with enabling the requirement for approvals from code owners in pull requests based on the `master` branch.

So far this additional setting in our protected branches wasn't necessary but as the company and the teams grow we need to enforce a more strict review process in changes in the style guide, making sure that key stakeholders are able to review modifications in order to minimize bugs and friction across the projects that depend on this style guide.