Use consistent_comma in literals instead of comma
Both settings enforce the use of trailing commas, but they behave
differently when multiple items are on one line.

With `comma`:

    [
      foo,
      bar,
    ]

    [
      one, three, five, seven, nine,
      two, four, six, eight
    ]                    # ^

With `consistent_comma`:

    [
      foo,
      bar,
    ]

    [
      one, three, five, seven, nine,
      two, four, six, eight,
    ]                    # ^

Sometimes it makes sense to group certain items if they are related, and
it seems weird that doing that would disable the trailing comma
convention.

Real world use case:

    # config/initializers/filter_parameter_logging.rb
    Rails.application.config.filter_parameters += [
      :password, :passw, :secret, :token, :_key, :crypt, :salt, :certificate, :otp, :ssn,
      # These exist on Twilio resources and can contain anything including PII, so we must treat them as unsafe.
      /task_?attributes/i, /worker_?attributes/i
    ]

The currently enforced style (`comma`) requires no trailing comma on
that last line, which is inconsistent with the style we have otherwise.