Disable EnforcedShorthandSyntax in Style/HashSyntax
We should not be forcing the shorthand syntax because of a few reasons:

1. That syntax doesn't always yields the best/clear code. It has
surprising behavior specially with method calls using the key names.
2. Ruby 3.1 shorthand syntax is still not supported by Sorbet.

We can fix 2, but not 1, so it is better to just allow people to chose
when to use the shorthand syntax in the context that makes sense.