Disallow pending cops
While we don't want to blindly enable all new cops, we should take a
stance on all cops, so our consumers don't have to.

Therefore, no cops should be marked as pending. When a Rubocop bump
introduces a new cop, we should triage it and either explicitly enable
or disable it.

If, for some reason, an update is required during a bad time for triage,
the updater should open an triage issue and mark the cop as disabled until it
is triaged.